Remove conflicting controllers with host app
Currently, while porting out functionality from the previous host app,
there are conflicts between stimulus controllers that are defined in the
host app and that the SelectComponent depends on as well. We don't keep
running both of these as then duplicate controllers will be running at
the same time (as we have namespaced the combobox-controllers) and the
behavior of the application could conflict.